+++ RELEASENOTES.md
@@ -22,6 +22,8 @@
         2000 ms respectively.
     *   Add `MediaExtractorCompat`, a new class that provides equivalent
         functionality to platform `MediaExtractor`.
+    *   Move `BasePreloadManager.Listener` to a top level
+        `PreloadManagerListener`.
 *   Transformer:
     *   Update parameters of `VideoFrameProcessor.registerInputStream` and
         `VideoFrameProcessor.Listener.onInputStreamRegistered` to use `Format`.